refactor: delete src/generated/types.ts
